% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut_groups}
\alias{cut_groups}
\title{Species -> group codes for one cut}
\usage{
cut_groups(x, phi, not_assigned = "0")
}
\arguments{
\item{x}{cocktail object}

\item{phi}{numeric cut (e.g., 0.3)}

\item{not_assigned}{label for species not captured}
}
\value{
tibble with species and group code
}
\description{
Species -> group codes for one cut
}
