% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktail_fuzzy.R
\name{cocktail_fuzzy}
\alias{cocktail_fuzzy}
\title{Fuzzy Cocktail memberships (no thresholding)}
\usage{
cocktail_fuzzy(x, vegmatrix, cuts = 0.3)
}
\arguments{
\item{x}{Result of \code{cocktail_cluster()}.}

\item{vegmatrix}{plots x species cover (numeric); used to get p/a for phi.}

\item{cuts}{numeric vector of phi cuts that define the parent groups (e.g., c(0.2,0.25,0.3)).}
}
\value{
data.frame with columns: cut, group, species, phi
}
\description{
Fuzzy Cocktail memberships (no thresholding)
}
